User auth and register

+++ base/views.py
@@ -1,9 +1,11 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
 
@@ -15,9 +17,14 @@ from .forms import RoomForm
 
 def loginPage(request):
 
+    page = 'Login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     # On récupère les valeurs du formulaire si ce dernier a bien une méthode POST
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         # On vérifie que l'utilisateur existe bien dans la base de donnée
@@ -34,12 +41,31 @@ def loginPage(request):
         else:
             messages.error(request, 'Username OR password does not match')
 
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+    context = {
+        'form': form
+    }
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+            
+    return render(request, 'base/login_register.html', context)
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -65,7 +91,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
-@login_required(login_url='login')
+@login_required(login_url='login') # Retreindre l'accès à la page
 def createRoom(request):
     form = RoomForm()
 
@@ -79,9 +105,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login') # Retreindre l'accès à la page
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here !')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -92,8 +122,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url='login') # Retreindre l'accès à la page
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here !')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
